필요없는 코드 삭제 및 코드 수정, 로그아웃 기능 수정

+++ src/main/java/com/example/sns1/post/PostService.java
@@ -51,8 +51,8 @@ public class PostService {
         return this.postRepository.save(post);
     }
 
-    public Post getPost(Integer id) {
-        Optional<Post> post = this.postRepository.findById(Long.valueOf(id));
+    public Post getPost(Long id) {
+        Optional<Post> post = this.postRepository.findById(id);
         
         if (post.isPresent()) {
             return post.get();
